% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.2side.ind.R
\name{mcmc.2side.ind}
\alias{mcmc.2side.ind}
\title{Run MCMC algorithm for the naive independence estimator.}
\usage{
mcmc.2side.ind(data, niter = 2400, nburn = 1200, nthin = 5, M = 200,
  inits = inits, proppars = list(lam01 = 0.05, lam02 = 0.05, sigma = 0.1, sx
  = 0.2, sy = 0.2), keepACs = TRUE, Rcpp = TRUE)
}
\arguments{
\item{data}{a list produced by sim2side or in the same format}

\item{niter}{number of MCMC iterations to run}

\item{nburn}{number of MCMC iterations to discard as burn in}

\item{nthin}{MCMC thinning parameter. Record output on every nthin iterations.  nthin=1 corresponds to no thinning}

\item{M}{The size of the augmented superpopulation}

\item{inits}{a list of user-supplied initial values.  inits=list(psi=psi,lam01=lam01,sigma=sigma)}

\item{proppars}{a list of tuning parameters for the proposal distributions}

\item{keepACs}{a logical indicating whether or not to return the posteriors for z, and s}
}
\value{
a list with the posteriors for the SCR parameters (out), s, z
}
\description{
This function fits the naive independence estimator from Augustine et al. 2016.  The data list should have the following elements:
1.  both, a n_both x 3 x K x J left side data array.  Not required if there are no double camera stations.
2.  left, a n_left x 3 x K x J left side data array.
3.  right, a n_right x 3 x K x J left side data array.
4. X a matrix with the X and Y trap locations in the first two columns and the number of cameras (1 or 2) at each trap in the third.
5. either buff or vertices.  buff is the fixed buffer for the traps to produce the state space.  It is applied to the minimum and maximum
X and Y locations, producing a square or rectangular state space.  vertices is a matrix with the X and Y coordinates of a polygonal state
space.
inits needs elements for psi1, psi2, and psi3 if there are some double camera stations or psi1 and psi2 if there are only single camera stations.
proppars needs elements for lam01 and lam02 if there are some double camera stations or lam01 if there are only single camera stations.
}
\author{
Ben Augustine, Andy Royle
}
