% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.SCRIPP.R
\name{mcmc.SCRIPP}
\alias{mcmc.SCRIPP}
\title{Run MCMC algorithm for SCR model with inhomogenous density and 1 categorical density covariate. Discrete state space.}
\usage{
mcmc.SCRIPP(data, niter = 2400, nburn = 1200, nthin = 5, M = 200,
  inits = inits, proppars = list(lam0 = 0.05, sigma = 0.1, sx = 0.2, sy =
  0.2, beta0 = 0.05, beta1 = 0.05), cellArea, Rcpp = TRUE)
}
\arguments{
\item{data}{a list produced by simSCRIPP or in the same format}

\item{niter}{number of MCMC iterations to run}

\item{nburn}{number of MCMC iterations to discard as burn in}

\item{nthin}{MCMC thinning parameter. Record output on every nthin iterations.  nthin=1 corresponds to no thinning}

\item{M}{The size of the augmented superpopulation}

\item{inits}{a list of user-supplied initial values.  inits=list(psi=psi,lam01=lam01,lam02=lam02,sigma=sigma)}

\item{proppars}{a list of tuning parameters for the proposal distributions}
}
\value{
a list with the posteriors for the SCR detection function and density covariate parameters (out), s, z
}
\description{
Run MCMC algorithm for SCR model with inhomogenous density and 1 categorical density covariate. Discrete state space.
}
\author{
Ben Augustine
}

