% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCR2DNAPPcheck.R
\name{SCR2DNAPPcheck}
\alias{SCR2DNAPPcheck}
\title{Calculate goodness of fit statistics for a model fit by SCR2DNAmcmc}
\usage{
SCR2DNAPPcheck(data, posterior, use, sharesig)
}
\arguments{
\item{data}{a list produced by simSCR2DNA or in the same format}

\item{posterior}{a posterior produced by SCR2DNAmcmc}

\item{use}{a vector if iteration values at which the statistics should be calculated}

\item{sharesig}{a logical indicating whether or not sigma was shared}
}
\description{
Calculate goodness of fit statistics for a model fit by SCR2DNAmcmc
}
\author{
Ben Augustine
}
