% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRmcmctf.R
\name{SCRmcmctf}
\alias{SCRmcmctf}
\title{Run MCMC algorithm for basic SCR model with a trap functionality file.}
\usage{
SCRmcmctf(data, niter = 2400, nburn = 1200, nthin = 5, M = 200,
  inits = inits, proppars = list(lam0 = 0.05, sigma = 0.1, sx = 0.2, sy =
  0.2))
}
\arguments{
\item{data}{a list produced by simSCR or in the same format}

\item{niter}{number of MCMC iterations to run}

\item{nburn}{number of MCMC iterations to discard as burn in}

\item{nthin}{MCMC thinning parameter. Record output on every nthin iterations.  nthin=1 corresponds to no thinning}

\item{M}{The size of the augmented superpopulation}

\item{inits}{a list of user-supplied initial values.  inits=list(psi=psi,lam0=lam0,sigma=sigma)}

\item{proppars}{a list of tuning parameters for the proposal distributions}
}
\value{
a list with the posteriors for the SCR parameters (out), s, z
}
\description{
This function runs the MCMC algorithm for the basic SCR model.  The data list should have the following elements:
1.  y, a n x J capture history
2.  X,  a matrix with the X and Y trap locations in the first two columns and the number of cameras (1 or 2) at each trap in the third.
3. either buff or vertices.  buff is the fixed buffer for the traps to produce the state space.  It is applied to the minimum and maximum
X and Y locations, producing a square or rectangular state space.  vertices is a matrix with the X and Y coordinates of a polygonal state
space.
4.  tf, a vector of length J containing the number of occasions each trap was operational
}
\author{
Ben Augustine, Andy Royle
}

