% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.conCatSMR.ID.R
\name{sim.conCatSMR.ID}
\alias{sim.conCatSMR.ID}
\title{Simulate data from the categorical spatial mark resight model}
\usage{
sim.conCatSMR.ID(N = 50, n.marked = 10, lam0 = 0.25, sigma = 0.5,
  K = 10, X = X, buff = 3, obstype = "bernoulli", ncat = ncat,
  pIDcat = pIDcat, gamma = gamma, IDcovs = IDcovs, pMarkID = c(1, 1),
  tlocs = 0, pID = 1, marktype = "premarked")
}
\description{
Will document after I defend. Should be able to get the idea in the example found in the mcmc.conCatSMR.ID help file.
}
\author{
Ben Augustine
}
