% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2side.plot.R
\name{sim2side.plot}
\alias{sim2side.plot}
\title{Plot trapping array and realized capture results of simulated data set}
\usage{
sim2side.plot(data, offset = 0.15, cexX = 1.5, cexB = 3, cexS = 3,
  cex0 = 1, cexT = 1, main = "", plotACs = TRUE, plottimes = 1)
}
\arguments{
\item{data}{a list produced by sim2side or in the same format}

\item{offset}{the amount to offset each camera for double traps}

\item{cexX}{the cex for representing the cameras}

\item{cexB}{the cex for complete identity activity centers}

\item{cexS}{the cex for partial identity activity centers}

\item{cex0}{the cex for uncaptured activity centers}

\item{cexT}{the cex for the text indicating B, L, R, or LR captures for each activity center}

\item{main}{the title for the plot}
}
\description{
This function plots the realized capture results from the simulation function. It plots the juxtaposition of
traps and activity centers, differentiating between complete identity individuals (B), partial identity individuals captured
on the left side only (L), right side only (R), left and right side but not simultanously (LR), and uncaptured individuals.
}
\examples{
\dontrun{
N=50
p01=0.13
p02=0.2
lam01=-log(1-p01)
lam02=-log(1-p02)
sigma=0.50
K=5
buff=2
xlim<- c(1,10)
ylim<- c(1,10)
X<- expand.grid(3:8,3:8) #6x6 trapping array
X=cbind(X,1) #add number of cameras at each trap
X[which(X[,2]\%in\%c(4,7)),3]=2 #switch the second and fifth row of traps to double cameras
#Simulate some data
data=sim2side(N=N,lam01=lam01,lam02=lam02,sigma=sigma,K=K,X=X,buff=buff)
sim2side.plot(data)
}
}
\author{
Ben Augustine
}

