% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2side.R
\name{sim2side}
\alias{sim2side}
\title{Simulate data from camera trap SCR study with partial identity}
\usage{
sim2side(N = 50, lam01 = 0.1, lam02 = 0.2, sigma = 0.5, K = 10,
  X = X, buff = 3, obstype = "bernoulli", tf = NA)
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam01}{the single side detection function hazard rate}

\item{lam02}{the double side detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the J x 3 matrix of trap locations and number of cameras at each trap. columns are X, Y, #cams (1 or 2)}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{tf}{an optional J x K matrix of dynamic trap operation indicating when and where 0, 1, or 2 cams were deployed}
}
\value{
a list containing the both, left, and right data sets, the both plus left only and both plus right
only data sets (stored in y.obs), the activity centers, the trap object, and several other data objects and summaries.
}
\description{
This function simulates data from a camera trap SCR study where some or all identities are partial.  The trap
matrix, X, should contain the X and Y locations of traps in columns 1 and 2 and the number of cameras at each
site in column 3 (either 1 or 2).  The extent of the state space is controlled by "buff", which buffers the
minimum and maximum X and Y extents.  Therefore, it probably only makes sense for square or rectangular grids.  Functionality
for user-provided polygonal state spaces will be added in the future.
}
\author{
Ben Augustine
}

