% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2sidetf.R
\name{sim2sidetf}
\alias{sim2sidetf}
\title{Simulate data from camera trap SCR study with partial identity and simulate a random trap failure process}
\usage{
sim2sidetf(N = 120, lam01 = 0.1, lam02 = 0.2, sigma = 0.5, K = 10,
  X = X, buff = 3, failprob = 0.05, faildur = 2, failtype = 2,
  obstype = "bernoulli")
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam01}{the single side detection function hazard rate}

\item{lam02}{the double side detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 3 matrix of trap locations and number of cameras at each trap. columns are X, Y, #cams}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{failprob}{the probability a camera or set of cameras at a site will fail on each occasion}

\item{faildur}{the number of occasions a camera remains inoperable}

\item{failtype}{1 if cameras at a site fail together and 2 if they fail independently}

\item{obstype}{observation type, either "bernoulli" or "poisson"}
}
\value{
a list containing the both, left, and right data sets, the both plus left only and both plus right
only data sets (stored in y.obs), the activity centers, the trap object, and several other data objects and summaries.
}
\description{
See the sim2side help file for a description of the capture process.  The camera failure process operates
like this:  cameras fail with probabilty failprob and remain disabled for faildur occasions.  If failtype=1, traps at
single sites fail together (bear attack?) and if failtype=2, they fail independently. This function was written to investigate
the importance of using the 2D trap operation file.
}
\author{
Ben Augustine
}
