% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCatSPIM.R
\name{simCatSPIM}
\alias{simCatSPIM}
\title{Simulate data from the categorical spatial partial identity model (SPIM)}
\usage{
simCatSPIM(N = 120, lam0 = 0.1, sigma = 0.5, K = 10, X = X,
  buff = 3, obstype = "bernoulli", ncat = ncat, pID = pID,
  gamma = gamma, IDcovs = IDcovs)
}
\description{
Will document after I defend. Should be able to get the idea in the example found in the mcmc.CatSPIM help file.
}
\author{
Ben Augustine
}
