% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simgenSPIM.R, R/simgenSPIMv2.R
\name{simgenSPIM}
\alias{simgenSPIM}
\alias{simgenSPIM}
\title{Simulate data with potentially partial individual indentities}
\usage{
simgenSPIM(N = 50, lam0 = 0.2, sigma = 0.5, K = 10, X = X, buff = 2,
  pID = 0.5, theta = 0.3, pconstrain = 0.3, obstype = "poisson")

simgenSPIM(N = 50, lam0 = 0.2, sigma = 0.5, K = 10, X = X, buff = 2,
  pID = 0.5, theta = 0.3, pconstrain = 0.3, obstype = "poisson")
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam0}{the detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{pID}{the probability that all of the samples from an individual will be perfectly identified}

\item{pconstrain}{the proportion of the identity constraint matrix that will have entry 1, allowing subcluster i and j to possibly
be the same individual}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{pgroup}{the proportion of partial identity samples that will be randomly grouped together in to correct identity subclusters}

\item{N}{a vector indicating the number of individuals to simulate}

\item{lam0}{the detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{pID}{the probability that all of the samples from an individual will be perfectly identified}

\item{pgroup}{the proportion of partial identity samples that will be randomly grouped together in to correct identity subclusters}

\item{pconstrain}{the proportion of the identity constraint matrix that will have entry 1, allowing subcluster i and j to possibly
be the same individual}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.

a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from a SCR study with potentially partial individual identities. If pID=1, normal SCR data
are produced, but if pID<1, some individuals will only be partially identified.  If pID=0, pgroup=0, and pconstrain=0, unidentified
 counts or detections (depending on observation model) will be produced as treated by Chandler and Royle (2013).  If pgroup>0, some
 singleton samples will be grouped back together.  This reproduces the scenario where you know some samples are the same individual,
 but other samples may belong to this focal individual as well.  We can also rule out matches between samples using a constraint
 matrix with elements 1 allowing matches and 0 disallowing them.  Because all the samples from the same individual must be allowed to 
 match with each other, a certain proportion of 1's are mandatory.  If pconstrain is greater than this proportion, random matches
 between subclusters are added (1's in the identity constraint matrix) until pconstrain percent of the identity constraint matrix
 are 1s.
 
The extent of the state space is controlled by "buff", which buffers the minimum and maximum X and Y extents.  Therefore, 
it probably only makes sense for square or rectangular grids.  Functionality for user-provided polygonal state spaces will 
be added in the future.

While the binomial observation model is included, I'm not sure it makes sense in practice unless it can be ensured
that multiple capture events from the same individual at the same trap on the same occasion never happen.  If you 
cannot fully identify all samples, you can't convert all trap-occasion counts>1 to a single encounter.

This function simulates data from a SCR study with potentially partial individual identities. If pID=1, normal SCR data
are produced, but if pID<1, some individuals will only be partially identified.  If pID=0, pgroup=0, and pconstrain=0, unidentified
 counts or detections (depending on observation model) will be produced as treated by Chandler and Royle (2013).  If pgroup>0, some
 singleton samples will be grouped back together.  This reproduces the scenario where you know some samples are the same individual,
 but other samples may belong to this focal individual as well.  We can also rule out matches between samples using a constraint
 matrix with elements 1 allowing matches and 0 disallowing them.  Because all the samples from the same individual must be allowed to 
 match with each other, a certain proportion of 1's are mandatory.  If pconstrain is greater than this proportion, random matches
 between subclusters are added (1's in the identity constraint matrix) until pconstrain percent of the identity constraint matrix
 are 1s.
 
The extent of the state space is controlled by "buff", which buffers the minimum and maximum X and Y extents.  Therefore, 
it probably only makes sense for square or rectangular grids.  Functionality for user-provided polygonal state spaces will 
be added in the future.

While the binomial observation model is included, I'm not sure it makes sense in practice unless it can be ensured
that multiple capture events from the same individual at the same trap on the same occasion never happen.  If you 
cannot fully identify all samples, you can't convert all trap-occasion counts>1 to a single encounter.
}
\author{
Ben Augustine

Ben Augustine
}
