% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOpenSPIM.R
\name{simOpenSPIM}
\alias{simOpenSPIM}
\title{Simulate data from open population SPIM.  Update help file later.  Add polygon SS later.}
\usage{
simOpenSPIM(N = c(40, 60, 80), gamma = NULL, phi = rep(0.8),
  lam01 = 0.2, lam02 = 0.2, sigma = 0.5, K = rep(10, 3), X = X,
  t = 3, M = M, sigma_t = NULL, buff = 3, obstype = "bernoulli",
  ACtype = "fixed", tf = NA)
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{tf}{an optional J x K matrix of dynamic trap operation indicating when and where 0, 1, or 2 cams were deployed}

\item{lam0}{the detection function hazard rate}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from a camera trap SCR study. The extent of the state space is controlled by "buff", which buffers the
minimum and maximum X and Y extents.  Therefore, it probably only makes sense for square or rectangular grids.  Functionality
for user-provided polygonal state spaces will be added in the future.
}
\author{
Ben Augustine
}
