% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSCR.R
\name{simSCR}
\alias{simSCR}
\title{Simulate data from a SCR study}
\usage{
simSCR(N = 120, lam0 = 0.2, lam0b = 0, sigma = 0.5, K = 10, X = X,
  buff = 3, obstype = "bernoulli")
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam0}{the detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{obstype}{observation type, either "bernoulli" or "poisson"}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from a camera trap SCR study. The extent of the state space is controlled by "buff", which buffers the
minimum and maximum X and Y extents.  Therefore, it probably only makes sense for square or rectangular grids.  Functionality
for user-provided polygonal state spaces will be added in the future.
}
\author{
Ben Augustine
}

