% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSCR2DNA.R
\name{simSCR2DNA}
\alias{simSCR2DNA}
\title{Simulate data from a SCR study with DNA collected from scat and hair snares (update help later)}
\usage{
simSCR2DNA(N = 120, lam01 = 0.2, lam01b = 0, lam02 = 0.2,
  sigma = c(0.5, 0.5), K1 = 10, K2 = 10, X1 = X1, X2 = X2, buff = 2)
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam01}{the hair snare detection function hazard rate}

\item{lam02}{the scat detection function hazard rate}

\item{sigma}{a vector of spatial scale parameters for the hair snare and scat detection functions (in that order)}

\item{K1}{the number of hair snare capture occasions}

\item{K2}{the number of scat capture occasions}

\item{X1}{the K1 x 2 matrix of hair snare locations}

\item{X2}{the K2 x 2 matrix of scat collection locations (descretized transects)}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{lam0b}{the hair snare detection function hazard rate after capture at a trap}
}
\value{
a list containing the capture histories (y1 for hair snares, y2 for scat collection), activity centers, trap objects, and several other data objects and summaries.
}
\description{
This function simulates data from a joint hair snare and scat collection SCR study. The extent of the state space is controlled by "buff", which buffers the
minimum and maximum X and Y extents.  Therefore, it probably only makes sense for square or rectangular grids.  Functionality
for user-provided polygonal state spaces will be added in the future.
}
\author{
Ben Augustine
}
