% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSCRhet.R
\name{simSCRhet}
\alias{simSCRhet}
\title{Simulate data from camera trap SCR study}
\usage{
simSCRhet(N = 120, lam0 = 0.2, sigma = 0.7, sig.lam0 = 0,
  sig.sigma = 0, sig.a0 = 0, beta = 1, K = 10, X = X, buff = 3,
  compensate = "Perfect", obstype = "bernoulli")
}
\arguments{
\item{N}{a vector indicating}

\item{b}{offset for trap-specific behavioral response}
}
\value{
a list containing
}
\description{
Simulate data from camera trap SCR study
}
\author{
Ben Augustine
}

