% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSCRIPP.R
\name{simSCRIPP}
\alias{simSCRIPP}
\title{Simulate data from camera trap SCR study with inhomogenous density in discrete space and one categorical covariate with 2 levels}
\usage{
simSCRIPP(N = 120, lam0 = 0.2, sigma = 0.5, K = 10, X = X, grid,
  Dparms, plot = TRUE)
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam0}{the detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{grid}{the data frame holding the state space points and covariate values. Should have columns labeled x, y, and cov with cov==1 or 2}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from a camera trap SCR study with inhomogenous density for two categorical covariates.
Binomial observation model.  Will generalize later.
}
\author{
Ben Augustine
}
