% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSCRtf.R
\name{simSCRtf}
\alias{simSCRtf}
\title{Simulate data from a SCR study and simulate a random trap failure process}
\usage{
simSCRtf(N = 120, lam0 = 0.2, lam0b = 0, sigma = 0.7, K = 10, X = X,
  buff = 3, failrate = 0.05, faildur = 2, obstype = "bernoulli")
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{lam0}{the detection function hazard rate}

\item{sigma}{the spatial scale parameter}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{faildur}{the number of occasions a trap remains inoperable}

\item{obstype}{observation type, either "bernoulli" or "poisson"}

\item{failprob}{the probability a trap will fail on each occasion}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
See the simSCR help file for a description of the capture process.  The trap failure process operates
like this:  traps fail with probabilty failprob and remain disabled for faildur occasions. This code does not keep up with
multiple cameras at a site like the 2side simulation code.  A trap is either on or off.
}
\author{
Ben Augustine
}

