% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSCR2DNA.R
\name{simSCR2DNA}
\alias{simSCR2DNA}
\title{Simulate data from a SCR study with DNA collected from scat and hair snares (update help later)}
\usage{
simSCR2DNA(N = 120, lam01 = 0.2, lam01b = 0, lam02 = 0.2, sigma = 0.5,
  K1 = 10, K2 = 10, X1 = X1, X2 = X2, buff = 3)
}
\arguments{
\item{N}{a vector indicating the number of individuals to simulate}

\item{sigma}{the spatial scale parameter}

\item{buff}{the distance to buffer the trapping array in the X and Y dimensions to produce the state space}

\item{lam0}{the detection function hazard rate}

\item{K}{the number of capture occasions}

\item{X}{the K x 2 matrix of trap locations}

\item{obstype}{observation type, either "bernoulli" or "poisson"}
}
\value{
a list containing the capture history, activity centers, trap object, and several other data objects and summaries.
}
\description{
This function simulates data from a camera trap SCR study. The extent of the state space is controlled by "buff", which buffers the
minimum and maximum X and Y extents.  Therefore, it probably only makes sense for square or rectangular grids.  Functionality
for user-provided polygonal state spaces will be added in the future.
}
\author{
Ben Augustine
}

